+++ src/Application/Query/Navigation/MenuViewQuery.php
@@ -10,12 +10,13 @@
 
 namespace Proximum\Vimeet\Application\Query\Navigation;
 
+use Proximum\Vimeet\Application\Query\Query;
 use Proximum\Vimeet\Domain\Model\Event;
 use Proximum\Vimeet\Domain\Model\Sheet;
 use Proximum\Vimeet\Domain\Model\StaticFormulation;
 use Proximum\Vimeet\Domain\Model\User;
 
-class MenuViewQuery
+class MenuViewQuery implements Query
 {
     /** @var Event */
     public $event;